# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardObjective = {
  title: string;
  detail: string;
};

export type DashboardSteps = {
  current: number;
  target: number;
};

export type DashboardHydration = {
  current: number;
  target: number;
};

export type DashboardBudget = {
  spent: number;
  total: number;
  currency: string;
};

export type DashboardInsight = {
  title: string;
  body: string;
  source: string;
};

export interface Readiness {
  score: number;
  label: string;
  note: string;
}

export interface WorkoutSummary {
  title: string;
  durationMinutes: number;
  completedExercises: number;
  totalExercises: number;
  status: string;
}

export interface NutritionSummary {
  calories: number;
  calorieTarget: number;
  protein: number;
  proteinTarget: number;
}

export interface Habit {
  id: string;
  label: string;
  completed: boolean;
}

export interface Dashboard {
  greeting: string;
  dateLabel: string;
  readiness: Readiness;
  objective: DashboardObjective;
  workout: WorkoutSummary;
  nutrition: NutritionSummary;
  steps: DashboardSteps;
  hydration: DashboardHydration;
  budget: DashboardBudget;
  habits: Habit[];
  insight: DashboardInsight;
}

export interface Exercise {
  id: string;
  name: string;
  order: number;
  targetWeight: number;
  targetSets: number;
  repRange: string;
  rir: number;
  restSeconds: number;
  previous: string;
  cue: string;
  completedSets: number;
}

export interface Workout {
  id: string;
  title: string;
  focus: string;
  durationMinutes: number;
  exercises: Exercise[];
}

export interface WorkoutSetInput {
  exerciseId: string;
  weight: number;
  reps: number;
  rir: number;
}

export interface SetLogResult {
  exerciseId: string;
  completedSets: number;
  message: string;
  nextAction: string;
}

export interface Meal {
  id: string;
  type: string;
  name: string;
  calories: number;
  protein: number;
  status: string;
}

export interface Budget {
  spent: number;
  total: number;
  remaining: number;
  currency: string;
}

export interface NutritionSnapshot {
  calories: number;
  calorieTarget: number;
  protein: number;
  proteinTarget: number;
  carbs: number;
  fat: number;
  meals: Meal[];
  budget: Budget;
}

export interface StrengthPoint {
  label: string;
  value: number;
  change: number;
}

export interface Mission {
  id: string;
  title: string;
  current: number;
  target: number;
  unit: string;
}

export interface Progress {
  weight: number;
  weightChange: number;
  consistency: number;
  streak: number;
  strength: StrengthPoint[];
  missions: Mission[];
}

export interface Profile {
  id: string;
  name: string;
  goal: string;
  targetWeight: number;
  targetDate: string;
  trainingDays: number;
  experience: string;
  equipment: string[];
}

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  goal?: string;
  targetWeight?: number;
  targetDate?: string;
  trainingDays?: number;
  experience?: string;
  equipment?: string[];
}