# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts/{workoutId}/sets:
    post:
      operationId: logWorkoutSet
      tags: [workouts]
      summary: Log a completed set
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSetInput"
      responses:
        "200":
          description: Logged set with coaching cue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetLogResult"
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get personal profile
      responses:
        "200":
          description: Personal profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update personal profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
components:
  schemas:
...
    WorkoutSetInput:
      type: object
      required: [exerciseId, weight, reps, rir]
      properties:
        exerciseId: { type: string }
        weight: { type: number }
        reps: { type: integer }
        rir: { type: integer }
    SetLogResult:
      type: object
      required: [exerciseId, completedSets, message, nextAction]
      properties:
        exerciseId: { type: string }
        completedSets: { type: integer }
        message: { type: string }
        nextAction: { type: string }
...
    Profile:
      type: object
      required: [id, name, goal, targetWeight, targetDate, trainingDays, experience, equipment]
      properties:
        id: { type: string }
        name: { type: string }
        goal: { type: string }
        targetWeight: { type: number }
        targetDate: { type: string }
        trainingDays: { type: integer }
        experience: { type: string }
        equipment:
          type: array
          items: { type: string }
    ProfileUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        goal: { type: string }
        targetWeight: { type: number }
        targetDate: { type: string }
        trainingDays: { type: integer }
        experience: { type: string }
        equipment:
          type: array
          items: { type: string }