# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  name: string;
  goal: string;
  targetWeight: number;
  targetDate: string;
  trainingDays: number;
  experience: string;
  equipment: string[];
}