# lib/api-zod/src/generated/types/setLogResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SetLogResult {
  exerciseId: string;
  completedSets: number;
  message: string;
  nextAction: string;
}